requirements.txt
# Core dependencies
telethon>=1.24.0
pandas>=1.3.0
python-dotenv>=0.19.0
asyncio-throttle>=1.0.2
tqdm>=4.62.0
aiohttp>=3.8.0
colorama>=0.4.4

# Optional dependencies for enhanced functionality
# Uncomment if needed:
# plotly>=5.0.0          # For data visualization
# jupyter>=1.0.0         # For interactive analysis
# openpyxl>=3.0.0        # For Excel file support
